voice_engine/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
faster-whisper>=1.0.0
piper-tts>=1.2.0
gTTS>=2.5.0
